fix(schemas): require integer credibility/relevance scores
Zod used z.number().min(1).max(10) for credibilityScore and relevanceScore,
accepting floats like 7.5, while the Gemini responseSchema declared
type: 'integer' for both. The Gemini gate was stricter than the Zod gate —
the inverse of the usual direction, and the same class of bug as the
original 'Mistrial most of the time' root cause (Zod and the Gemini schema
disagreeing on a numeric constraint). Currently masked because the integer
gate happens to fire first at the API. Tighten Zod to .int().min(1).max(10)
so the two gates agree and a float fails fast on the client too, matching
the existing amount field at gameSchemas.ts:116.

+++ src/schemas/gameSchemas.ts
@@ -284,7 +284,7 @@ export const WitnessSchema = z.strictObject({
   role: WitnessRoleEnum,
   bias: BiasIndicatorEnum,
   statement: noJury(z.string().max(1000)).describe("A summary of their expected testimony."),
-  credibilityScore: z.number().min(1).max(10),
+  credibilityScore: z.number().int().min(1).max(10),
   // Voiced testimony beats for the trial phase. Which side conducts direct
   // (and which crosses) is derived from `bias` — PROSECUTION/NEUTRAL
   // witnesses are the People's, DEFENSE witnesses are the defense's.
@@ -302,7 +302,7 @@ export const EvidenceSchema = z.strictObject({
   // counsel-voiced — the full detail stays hidden until the exhibit is
   // actually offered.
   disclosureSummary: noJury(z.string().min(1).max(400)).describe("Counsel's brief, unverified summary of the item as disclosed in discovery, spoken to the court."),
-  relevanceScore: z.number().min(1).max(10).describe("Scale of 1-10 on impact to the case."),
+  relevanceScore: z.number().int().min(1).max(10).describe("Scale of 1-10 on impact to the case."),
   objectionRisk: z.enum(['LOW', 'MEDIUM', 'HIGH']).describe("Likelihood of opposing counsel objecting."),
   targetElementId: z.string().min(1).max(40).nullable().describe("The ID of the StatuteElement this evidence is meant to prove."),
   isAdmitted: z.boolean().optional().transform((): boolean => false).describe("Always initialized to false. Mutated by player action during the trial phase."),
